# project-root/asm-msa-identity/src/main/java/com/company/asm/identity/service/dto/MessageResponseDto.java
package com.company.asm.identity.service.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class MessageResponseDto {
    private String message;
}